Adds contrallers folder and moves Joi validation to modelContacts
