+++ controllers/contacts/getAll.js
@@ -1,0 +1,14 @@
+const { Contact } = require('../../models/modelContacts');
+
+const listContacts = async () => {
+  try {
+    const contacts = await Contact.find({});
+    return contacts;
+  } catch (error) {
+    console.log(error);
+  } 
+}
+
+module.exports = {
+  listContacts,
+}
