+++ routes/api/contacts.js
@@ -1,6 +1,6 @@
 const express = require('express');
-const Joi = require('joi');
-const { listContacts, getContactById, removeContact, addContact, updateContact, updateStatus} = require('../../models/contacts');
+const { addContactValidation, updateContactValidation, updateStatusValidation } = require('../../models/modelContacts');
+const { listContacts, getContactById, removeContact, addContact, updateContact, updateStatus} = require("../../controllers/contacts/index");
 const router = express.Router();
 
 router.get('/', async (req, res, next) => {
@@ -34,21 +34,7 @@ router.post('/', async (req, res, next) => {
     return res.status(400).json({ message: `Missing required phone field` });
   }
 
-  const schema = Joi.object({
-    name: Joi.string()
-      .pattern(/^\w+\s/)
-      .min(3)
-      .max(30)
-      .required(),
-    email:Joi.string()
-      .email({ minDomainSegments: 2, tlds: { allow: ['com', 'net', 'org'] } })
-      .required(),
-    phone: Joi.string()
-      .pattern(/^(?:\+38)?(?:\(\d{3}\)[ .-]?[0-9]{3}[ .-]?[0-9]{2}[ .-]?[0-9]{2}|044[ .-]?[0-9]{3}[ .-]?[0-9]{2}[ .-]?[0-9]{2}|044[0-9]{7})/)
-      .required()
-  });
-
-  const validation = schema.validate(req.body); 
+  const validation = addContactValidation(req.body)
   if (validation.error) {
     return res.status(400).json({ message: `Failed because ${validation.error}` });
   }
@@ -59,10 +45,9 @@ router.post('/', async (req, res, next) => {
 
 router.delete('/:contactId', async (req, res, next) => {
   
-  const prevContacts = await removeContact(req.params.contactId);
-  const idArr = prevContacts.map(contact => contact.id);
-
-  if (idArr.includes(req.params.contactId)) {
+  const contact = await removeContact(req.params.contactId);
+  
+  if (contact) {
     return res.status(200).json({ message: 'Contact deleted' });
   } 
 
@@ -77,21 +62,7 @@ router.put('/:contactId', async (req, res, next) => {
     return res.status(400).json({ message: 'Missing fields' });
   }
 
-  const schema = Joi.object({
-    name: Joi.string()
-      .pattern(/^\w+\s/)
-      .min(3)
-      .max(30)
-      .optional(),
-    email: Joi.string()
-      .email({ minDomainSegments: 2, tlds: { allow: ['com', 'net', 'org'] } })
-      .optional(),
-    phone: Joi.string()
-      .pattern(/^(?:\+38)?(?:\(\d{3}\)[ .-]?[0-9]{3}[ .-]?[0-9]{2}[ .-]?[0-9]{2}|044[ .-]?[0-9]{3}[ .-]?[0-9]{2}[ .-]?[0-9]{2}|044[0-9]{7})/)
-      .optional(),
-  });
-
-  const validation = schema.validate(req.body);
+  const validation = updateContactValidation(req.body);
   
   if (validation.error) {
     return res.status(400).json({ message: `Failed because ${validation.error}` });
@@ -107,12 +78,7 @@ router.put('/:contactId', async (req, res, next) => {
 })
 
 router.patch('/:contactId/favorite', async (req, res, next) => {
-  const schema = Joi.object({
-    favorite: Joi.boolean()
-      .required(),
-  })
-
-  const validation = schema.validate(req.body); 
+  const validation = updateStatusValidation(req.body); 
 
    if (validation.error) {
     return res.status(400).json({ message: `Failed because ${validation.error}` });
